+++ kiplot/out_kibom.py
@@ -21,15 +21,35 @@ class KiBoM(BaseOutput):  # noqa: F821
         self._sch_related = True
         # Options
         with document:
+            self.number = 1
+            """ Number of boards to build (components multiplier) """
+            self.variant = ''
+            """ Board variant(s), used to determine which components
+                are output to the BoM. To specify multiple variants,
+                with a BOM file exported for each variant, separate
+                variants with the ';' (semicolon) character """
+            self.conf = 'bom.ini'
+            """ BoM configuration file, relative to PCB """
+            self.separator = ','
+            """ CSV Separator """
             self.format = 'HTML'
             """ [HTML,CSV] format for the BoM """  # pragma: no cover
 
     def run(self, output_dir, board):
         check_script(CMD_KIBOM, URL_KIBOM)
         format = self.format.lower()
-        prj = os.path.splitext(os.path.relpath(GS.pcb_file))[0]
-        logger.debug('Doing BoM, format '+format+' prj: '+prj)
-        cmd = [CMD_KIBOM, prj+'.xml', os.path.join(output_dir, os.path.basename(prj))+'.'+format]
+        prj = os.path.splitext(os.path.abspath(GS.pcb_file))[0]
+        config = os.path.join(os.path.dirname(os.path.abspath(GS.pcb_file)), self.conf)
+        logger.debug('Doing BoM, format {} prj: {} config: {}'.format(format, prj, config))
+        cmd = [CMD_KIBOM,
+               '-n', str(self.number),
+               '--cfg', config,
+               '-s', self.separator]
+        if GS.debug_enabled:
+            cmd.append('-v')
+        if self.variant:
+            cmd.extend(['-r', self.variant])
+        cmd.extend([prj+'.xml', os.path.join(output_dir, os.path.basename(prj))+'.'+format])
         logger.debug('Running: '+str(cmd))
         try:
             cmd_output = check_output(cmd, stderr=STDOUT)